fix: upload read directory files by DirectoryReader limit maximum files
