+++ components/Upload/style/index.less
@@ -229,7 +229,7 @@
     }
   }
 
-  &-item-done &-item-text-file-icon {
+  &-item-done &-file-icon {
     color: @upload-text-item-color-file-icon_success;
   }
