+++ components/Upload/style/token.less
@@ -21,6 +21,7 @@
 
 @upload-drag-font-size: @font-size-body-3;
 @upload-drag-border-radius: @radius-small;
+@upload-drag-tip-margin-top: 0;
 @upload-drag-color-text: var(--color-text-1);
 @upload-drag-border-style: @border-dashed;
 @upload-drag-border-width: @border-1;
@@ -51,7 +52,7 @@
 @upload-text-item-margin-remove-icon-left: @spacing-6;
 @upload-text-item-color-remove-icon: var(--color-text-2);
 @upload-text-item-color-status-icon: var(--color-white);
-@upload-text-item-color-file-icon_success: @color-success-6;
+@upload-text-item-color-file-icon_success: @color-primary-6;
 @upload-text-item-color-progress-bg_hover: rgba(var(--gray-10), 0.2);
 @upload-text-item-color-progress-bg_hover_active: @color-primary-7;
 @upload-text-item-size-file-icon: @size-4;
